Get the astra driver to exit if it can't find a device
Without this change the driver would switch to finding ANY device, which
meant that if one camera had a wrong ID (or more commonly the hardware
didn't come up), it could steal another camera. So we could end up with
the "frontcenter" node owning the proximity_front camera which effectively
makes both of them useless!

Another (bad) possiblity is that the proximity_front camera could claim
some other camera and then we would try to drive with _very_ invalid data.

In addition to fixing the assignment problem described above, this change,
when combined with a change to the launch file to make the driver a
"required" nodelet, will allow systemd to manage the shutdown, restart,
recovery of the cameras individually.