astra_driver: don't leak fds or use when invalid
While this code technically did check the return code of open it would
go ahead and use it anyway even if it failed. Additionally if it
succeeded it would never close the file descriptor. This means it would
leak the file descriptor. In the crazy case the unlock failed for some
reason closing the file descriptor would also work to unlock it since
locks only apply to open file descriptors.