astra_camera: add signal handler to allow clean shutdown
When this node/nodelet is stopped it can sometimes leave the camera is a
bad state where the only solution is to restart the camera itself. The
only way to restart the camera is to restart the whole system or
phsyically unplug the camera. Since both of those solutions are
impractical a solution that prevents us from ever entering that state is
needed.

By adding a signal handler and telling ros to shutdown we should allow
the camera to get shutdown more cleanly and less abrubtly. Testing shows
that this seems to reduce the number of times that we get the camera
into a bad state.

We handle both SIGINT and SIGTERM because the nodelet will sent SIGINT
first followed by SIGTERM. This should give us enough time to cleanly
shut everything down. If both of these timeout it will eventually send
SIGKILL, which we can't handle, and we will not get to shutdown cleanly.
We generally do have enough time to shutdown though before that happens.

It would be great if the underlying issue with the devices was fixed but
this should help for now.