add optional ability to use lock file during camera init
While the upstream code has an optional way to synchronize bringup
using a shared memory segment and a semaphore, this mechanism is
brittle, and only designed to work once, as it does not handle
shutdown and restart of only one of the cameras. It also requires
knowing the number of cameras a priori in a parameter and requires
more coordination than a simple global advisory lock file. Finally,
this semaphore mechanism does not protect device enumeration, which
should not be done concurrently as it opens each device to read
serial numbers. The advisory lock is simpler, and more robust, as on
any process exit it will be freed by the operating system. Also this
advisory lock system has years of runtime at Badger, so port it even
though there is an upstream startup synchronization.