badger: add dynamic reconfigure for laser, LDP on/off
Add dynamic reconfigure parameters for laser emitter and proximity
sensor (LDP). Default to laser emitter enabled, LDP disabled, as LDP
on our robots is obscured enough by covers to trip intermittently.
Attempting to disable LDP on non-supported devices results in errors
and, on the Astra Mini, even breaks the RGB stream, so only attempt
to set the LDP state from dynamic reconfigure for Stereo S U3
devices.

Also, starting the streams auto-enables the emitter, so it set the
emitter state after the streams are enabled.

Set the LDP state when the streams are disabled, as it seems that
setting it when enabled also can cause problems.