badger: fix default intrinsics, taking SXGA into account
Use height for calculating default intrinsics instead of 3/4 of the
width (which is only true for 4:3 video modes).

Adjust the height for SXGA (1280x1024) which is an oddball.
SXGA is only useful on 4:3 devices, and in such cases keep the default
info consistent by assuming the images always align at the top corner.
In reality some devices (at least the Mini) run natively in 1280x1024 and
the other modes chop off the bottom data, which means the default center
point is not correct, and should really be the center point of SXGA
adjusted. However, for legacy reasons, pretend that the default center
point of the 4:3 modes is correct, as adjusting that assumption now would
ruin existing extrensics calibrations which used the faulty intrensic
center point. This only affects devices which do not store true intrensics
anyway, which are legacy devices such as the Mini.