badger: re-enable Stereo S support (HACK)
From inspection, 0x00103E30 is the location of the supported USB
vendor ids and product ids in binary (for this compilation). The
first entry was for product id '0x0300' which is not listed in any
udev rules or documents, so likely represents some early hardware we
will never see. Alter this entry's pid to '0x0608' (the Stereo S's
pid) to add back Stereo S support to liborbbec.so. This at least
allows the Stereo S to enumerate. It is unknown at the time of
writing this commit if this successfully enables all features of the
Stereo S.

For future rebases, if it is necessary to find this table again, just
do a binary search for 'c5 2b 04 04' (the 16-bit numbers are little
endian). The table is found if the other binary values in this region
also represent the Orbbec vendor id (0x2bc5) and the other product
ids in numerical order.