badger: let dyn config callback mutate config
The dynamic reconfigure interface allows the callback to mutate the
passed in configuration on purpose so it can prevent setting invalid
combinations of parameters for specific scenarios that can not be
handled by the ranges in the .cfg file. Remove the const-ness of the
internal interface from the AstraConfig reference to allow mutation.