correct SXGA for Astra Mini Pro
The mini pro supports 1280x1024, but intrensics are stored in 4:3 for
VGA (640x480). Scaling isn't sufficient to fix the central point when
in 1280x1024. By inspection, the 1280x920 image centers properly on
the 1280x1024 image (so 1280x920 is cutting off half the top and
bottom of the 1280x1024 image). If the mini pro is in 1280x1024 mode,
add (1024-920)/2 = 52 to the cy points.