fixup: update SXGA handling for astra mini.
Only propegate broken legacy default camera info on the mini and not
mini pro.

For the Astra Mini, adjust the height for SXGA (1280x1024). On the Astra
Mini, the other video modes cut off the raw data from 1280x1024.
However, due to legacy reasons, pretend the VGA mode intrinsics are
correct and use only 960 as the effective height when calculating the
central y point. The legacy reasons are that if the central point were
accurately moved down to the proper spot for all modes, it would
invalidate other parameters and existing calibrations, which would
require all astra mini calibrations to be re-run. Instead, just keep the
broken old legacy behaior. The Astra Mini Pro doesn't need any extra
correction here, as the non 1280x1024 modes are centered on 1280x1024,
so the half-height point is the correct default for all modes on Astra
Mini Pro.