badger: alter focal length, central point for mini's
The astra mini default focal length is wrong. Scale the focal length
by .91. This number was obtained from adjusting the mask during
fleet-wide tuning of fidicual free spoiler detection.

Also, due to legacy reasons, the y central point needs to be adjusted
up (subtracted) to compensate for the error caused by the incorrect
SXGA intrinsics (because extrinsics are calibrated wrongly). This is
totally bogus, but all existing Astra Mini calibrations depend on
this bogusness and it is too much work to recalibrate them all.
Fixing this issue here allows users of RGB intriniscs (such as
docking) to not need to fixup the broken intrinsics (so such code can
work with newer types of cameras which are not broken).