Fix modify-after-publish bug
The camera driver was modifying the time stamp on the same camera_info
object each time it published an image via image_transport. This means
that, if it takes more than one full frame period for the message to get
serialized to external subscribers, it is possible for the next
iteration to modify the timestamp before it went out.  This was observed
when Rviz subscribed to a throttled image topic over wifi. It could
cause the cameara_infos pulished by the throttled node to be off by one
image period, so that they never match exactly with the timestamp in the
image messages so that image_transport was would become unsynchronized
for throttled images. This could cause the auto-floor-leveling to never
work, since it uses throttled topics, which could prevent floor
calibration from working, since it waits for the auto-floor-leveler.

ROS-1531 ("Unable to do depthcam floor calibration due to image_transport unsynchronized")
ROS-376 ("Front proximity depthcam placement -> calibration issue")